fix clippy (#1725)
* fix clippy

* fix clippy fastfield codecs

* fix clippy bitpacker

* fix clippy common

* fix clippy stacker

* fix clippy sstable

* fmt

+++ fastfield_codecs/src/compact_space/build_compact_space.rs
@@ -208,7 +208,7 @@ impl CompactSpaceBuilder {
             };
             let covered_range_len = range_mapping.range_length();
             ranges_mapping.push(range_mapping);
-            compact_start += covered_range_len as u64;
+            compact_start += covered_range_len;
         }
         // println!("num ranges {}", ranges_mapping.len());
         CompactSpace { ranges_mapping }
